+++ src/pint/models/binary_ell1.py
@@ -16,7 +16,6 @@ from pint.models.stand_alone_psr_binaries import binary_orbits as bo
 from pint.models.stand_alone_psr_binaries.ELL1_model import ELL1model
 from pint.models.stand_alone_psr_binaries.ELL1H_model import ELL1Hmodel
 from pint.models.stand_alone_psr_binaries.ELL1k_model import ELL1kmodel
-from pint.models.stand_alone_psr_binaries.ELL1plus_model import ELL1plusmodel
 from pint.models.timing_model import MissingParameter
 from pint.utils import taylor_horner_deriv
 from pint import Tsun
@@ -61,7 +60,7 @@ class BinaryELL1(PulsarBinary):
 
     This binary model uses a rectangular representation for the eccentricity of an orbit,
     resolving complexities that arise with periastron-based parameters in nearly-circular
-    orbits. It also makes certain approximations that are invalid when the eccentricity
+    orbits. It also makes certain approximations (up to O(e^2)) that are invalid when the eccentricity
     is "large"; what qualifies as "large" depends on your data quality. A formula exists
     to determine when the approximations this model makes are sufficiently accurate.
 
@@ -79,8 +78,22 @@ class BinaryELL1(PulsarBinary):
     References
     ----------
     - Lange et al. (2001), MNRAS, 326 (1), 274–282 [1]_
+    - Zhu et al. (2019), MNRAS, 482 (3), 3249-3260 [2]_
 
-    .. [1] https://ui.adsabs.harvard.edu/abs/2001MNRAS.326..274L/abstract
+    .. [1] https://ui.adsabs.harvard.edu/abs/2019MNRAS.482.3249Z/abstract
+    .. [2] https://ui.adsabs.harvard.edu/abs/2001MNRAS.326..274L/abstract
+
+
+    Notes
+    -----
+    This includes o(e^2) expression for Roemer delay from Norbert Wex and Weiwei Zhu
+    This is equaiton (1) of Zhu et al (2019) but with a corrected typo:
+        In the first line of that equation, ex->e1 and ey->e2
+        In the other lines, ex->e2 and ey->e1
+    See Email from Norbert and Weiwei to David on 2019-Aug-08
+    The dre expression comes from Norbert and Weiwei; the derivatives
+    were calculated by hand for PINT
+
     """
 
     register = True
@@ -285,20 +298,7 @@ class BinaryELL1(PulsarBinary):
 
 
 class BinaryELL1H(BinaryELL1):
-    """
-    ELL1 modified to use ``H3``, ``H4``, and ``STIGMA`` parameters, from Freire and Wex (2010).
-
-    The :class:`~pint.models.binary_ell1.BinaryELL1H` model parameterizes the Shapiro
-    delay differently compare to the :class:`~pint.models.binary_ell1.BinaryELL1`
-    model. A fourier series expansion is used for the Shapiro delay:
-
-    .. math::
-
-        \\Delta_S = -2r \\left( \\frac{a_0}{2} + \\Sum_k (a_k \\cos k\\phi + b_k \\sin k \phi) \\right)
-
-    The first two harmonics are generlly absorbed by the ELL1 Roemer delay.
-    Thus, :class:`~pint.models.binary_ell1.BinaryELL1H` uses the series from the third
-    harmonic and higher.
+    """ELL1 modified to use H3 parameter for Shapiro delay.
 
     The actual calculations for this are done in
     :class:`pint.models.stand_alone_psr_binaries.ELL1_model.ELL1model`.
@@ -308,17 +308,15 @@ class BinaryELL1H(BinaryELL1):
     .. paramtable::
         :class: pint.models.binary_ell1.BinaryELL1H
 
-    In addition, if ``H3`` and one of ``H4`` or ``STIGMA`` is set, you can access the inferred ``SINI`` and ``M2``
-    values via the ``.sini`` and ``.m2`` properties
-
+    Note
+    ----
+    Only the Medium-inclination case model is implemented.
 
     References
     ----------
-    - Freire and Wex (2010), MNRAS, 409, 199 [1]_
+    - Freire & Wex (2010), MNRAS, 409 (1), 199-212 [1]_
 
     .. [1] https://ui.adsabs.harvard.edu/abs/2010MNRAS.409..199F/abstract
-
-     for Shapiro delay.
     """
 
     register = True
@@ -376,13 +374,12 @@ class BinaryELL1H(BinaryELL1):
         if self.H4.quantity is not None:
             self.binary_instance.fit_params = ["H3", "H4"]
             # If have H4 or STIGMA, choose 7th order harmonics
-            if self.NHARMS.value < 7:
-                self.NHARMS.value = 7
+            self.NHARMS.value = max(self.NHARMS.value, 7)
             if self.STIGMA.quantity is not None:
                 raise ValueError("ELL1H can use H4 or STIGMA but not both")
-        elif self.STIGMA.quantity is not None:
+
+        if self.STIGMA.quantity is not None:
             self.binary_instance.fit_params = ["H3", "STIGMA"]
-            log.debug("Using exact ELL1H Shapiro delay for STIGMA > 0")
             self.binary_instance.ds_func = self.binary_instance.delayS_H3_STIGMA_exact
             if self.STIGMA.quantity <= 0:
                 raise ValueError("STIGMA must be greater than zero.")
@@ -394,36 +391,9 @@ class BinaryELL1H(BinaryELL1):
         # if self.H3.quantity is None:
         #     raise MissingParameter("ELL1H", "H3", "'H3' is required for ELL1H model")
 
-    @property
-    def sini(self):
-        if self.H4.quantity is not None:
-            return (
-                2
-                * self.H3.quantity
-                * self.H4.quantity
-                / (self.H3.quantity**2 + self.H4.quantity**2)
-            )
-        elif self.STIGMA.quantity is not None:
-            return 2 * self.STIGMA.quantity / (1 + self.STIGMA.quantity**2)
-        else:
-            return None
-
-    @property
-    def m2(self):
-        if self.H4.quantity is not None:
-            return (
-                (self.H3.quantity) ** 4 / (self.H4.quantity) ** 3 / Tsun
-            ).decompose() * u.Msun
-        elif self.STIGMA.quantity is not None:
-            return (
-                (self.H3.quantity / self.STIGMA.quantity**3) / Tsun
-            ).decompose() * u.Msun
-        else:
-            return None
-
 
 class BinaryELL1k(BinaryELL1):
-    """ELL1k binary model modified for short-orbital period binaries, from Susobhanan et al. (2018)
+    """ELL1k binary model.
 
     Modified version of the ELL1 model applicable to short-orbital period binaries where
     the periastron advance timescale is comparable to the data span. In such cases, the
@@ -457,9 +427,6 @@ class BinaryELL1k(BinaryELL1):
         self.binary_model_name = "ELL1k"
         self.binary_model_class = ELL1kmodel
 
-        # remove the funcparameter, but it back for real
-        self.remove_param("OMDOT")
-
         self.add_param(
             floatParameter(
                 name="OMDOT",
@@ -480,7 +447,6 @@ class BinaryELL1k(BinaryELL1):
 
         self.remove_param("EPS1DOT")
         self.remove_param("EPS2DOT")
-        self.remove_param("EDOT")
 
     def validate(self):
         """Validate parameters."""
@@ -512,49 +478,3 @@ class BinaryELL1k(BinaryELL1):
             self.EPS2.quantity = (1 + lnedot * dt) * (
                 eps20 * np.cos(omdot * dt) - eps10 * np.sin(omdot * dt)
             )
-
-
-class BinaryELL1plus(BinaryELL1):
-    """ELL1+ binary model for higher-order eccentricity correction
-
-    Modified version of the ELL1 model applicable to long-period binaries where
-    x*ecc**2 may be larger than RMS/sqrt(Ntoa), but x*ecc**3 will be smaller.
-
-    The actual calculations for this are done in
-    :class:`pint.models.stand_alone_psr_binaries.ELL1plus_model.ELL1plusmodel`.
-
-    It supports all the parameters defined in :class:`pint.models.pulsar_binary.PulsarBinary`.
-
-    Parameters supported:
-
-    .. paramtable::
-        :class: pint.models.binary_ell1.BinaryELL1plus
-
-    References
-    ----------
-    - Zhu et al. (2019), MNRAS, 482 (3), 3249-3260 [1]_
-
-    .. [1] https://ui.adsabs.harvard.edu/abs/2019MNRAS.482.3249Z/abstract
-
-
-    Notes
-    -----
-    o(e^2) expression for Roemer delay from Norbert Wex and Weiwei Zhu
-    This is equaiton (1) of Zhu et al (2019) but with a corrected typo:
-        In the first line of that equation, ex->e1 and ey->e2
-        In the other lines, ex->e2 and ey->e1
-    See Email from Norbert and Weiwei to David on 2019-Aug-08
-    The dre expression comes from Norbert and Weiwei; the derivatives
-    were calculated by hand for PINT
-    """
-
-    register = True
-
-    def __init__(self):
-        super().__init__()
-        self.binary_model_name = "ELL1+"
-        self.binary_model_class = ELL1plusmodel
-
-    def validate(self):
-        """Validate parameters."""
-        super().validate()
